feat: add cluster and implement create server fn
